requirements.txt
Django>=4.2
matplotlib
cohere
python-dotenv
